LayoutTests/media/auto-play-in-sandbox-with-allow-scripts.html

<!DOCTYPE html>
<title>Test that play event fires when "src" set with an autoplay attribute in a sandbox with allows-scripts.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<iframe
    sandbox="allow-scripts allow-same-origin"
    src="resources/auto-play-in-sandbox-with-allow-scripts-iframe.html">
</iframe>
<script>
async_test(function(t) {
    var iframe = document.querySelector("iframe");

    iframe.onload = t.step_func(function() {
        var video = iframe.contentDocument.querySelector("video");
        assert_true(video.paused);
        video.onplay = t.step_func_done();
        video.src = "content/test.mp4";
    });
});
</script>
